added a data class to hold a bunch of data. Meso is made of weeks. weeks are made of days
