+++ Day.java
@@ -4,4 +4,8 @@ public class Day{
   public Day(){
     
   }
+
+  public void print(){
+    System.out.println("day");
+  }
 }
