+++ Exercise.java
@@ -2,8 +2,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Exercise {
-  public static final int NUM_BODY_PARTS = 11; // Number of muscle groups
-  public static final String[] BODY_PARTS = {"abs", "calves", "biceps", "triceps", "chest", "frontDelts", "sideDelts", "rearDelts", "quads", "hamstrings", "glutes", "lats", "rhomboids", "traps", "SpineErectors"};
 
 
   private String name;
