+++ Main.java
@@ -1,7 +1,7 @@
 class Main {
   public static void main(String[] args) {
-           Mesocycle test = new Mesocycle();
-        test.print();
+    Mesocycle test = new Mesocycle();
+    test.print();
 
         //will generate a split that brings each muscle group from mev to mrv over a mesocycle (not including deload)
         // add a data type that represents the amount of sets the user accumulates for each muscle group (display this to the user)
