+++ Mesocycle.java
@@ -6,11 +6,15 @@ public class Mesocycle{
   public Mesocycle(){//take user input as parameter
     numWeeks=8;//make user input # of weeks
     meso = new Week[numWeeks];
+    for(int i=0; i<numWeeks; i++){
+      meso[i]=new Week();
+    }
+    fillWeeks();
   }
 
   private void fillWeeks(){
     for(int i=0; i<numWeeks; i++){
-      //meso[i].ConstructWeekObject();//fill this out in Week class
+      meso[i].ConstructWeekObject();//fill this out in Week class
     }
   }
